requirements.txt
black==19.3b0
dj-database-url==0.5.0
django-environ==0.4.5
django-extensions==2.2.3
django-filter==2.1.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
Django==2.2
djangorestframework==3.9.2
psycopg2==2.7.7
uWSGI==2.0.18
Werkzeug==0.16.0
